Tests/LibWeb/Text/input/HTML/iframe-element-loading.html

<iframe id="testFrame"></iframe>
<script src="../include.js"></script>
<script>
    test(() => {
        const iframe = document.getElementById('testFrame');

        println(`loading = ${iframe.loading}`);
        println('');

        iframe.loading = 'lazy';

        println(`loading = ${iframe.loading}`);
        println('');

        iframe.loading = 'invalid-value';
        println(`loading = ${iframe.loading}`);
    });
</script>
