Tests/LibWeb/Text/expected/HTML/iframe-element-loading.txt

   loading = eager

loading = lazy

loading = eager
